ds 07: rebuild home page into src/features/home per figma board 17830:75689
why: KR3 architecture move (thin page + flow hook + dumb views, matching
features/payments/flows/semantic-request) and the DS 07 home layout: bare
avatar top-left, rewards link top-right, centered balance with
add/send/request submenu, cta card slot composed as-is, activity feed on
the ListItem primitive.

- home/page.tsx is now a shim; all logic in src/features/home/
- balance visibility moved to a useSyncExternalStore over the persisted
  user preference so page/flow hook stay useState-free
- modal priority chain extracted unchanged into HomeModals
- TransactionCard, KycStatusItem, BadgeStatusItem, CardUnlockHistoryItem
  now render through the ListItem primitive (shared with /history)
- ListItem + Global/Card gain aria-label passthrough (replaces the
  CardUnlockHistoryItem button wrapper)
- withdraw quick action dropped from home per the figma board (3 submenu
  buttons) — flagged in the PR as a product-affecting change
