+++ src/app/(mobile-ui)/home/page.tsx
@@ -1,500 +1,7 @@
 'use client'
 
-import { Button, type ButtonSize, type ButtonVariant } from '@/components/0_Bruddle/Button'
-import PageContainer from '@/components/0_Bruddle/PageContainer'
-import { Icon } from '@/components/Global/Icons/Icon'
-import Loading from '@/components/Global/Loading'
-import HomeHistory from '@/components/Home/HomeHistory'
-import { UserHeader } from '@/components/UserHeader'
-import { useAuth } from '@/context/authContext'
-import { useWallet } from '@/hooks/wallet/useWallet'
-import { useUserStore } from '@/redux/hooks'
-import { formatExtendedNumber, getUserPreferences, updateUserPreferences } from '@/utils/general.utils'
-import { printableUsdc } from '@/utils/balance.utils'
-import { useDisconnect } from 'wagmi'
-import Link from 'next/link'
-import { useEffect, useMemo, useState, useCallback, lazy, Suspense } from 'react'
-import { twMerge } from 'tailwind-merge'
-import { useAccount } from 'wagmi'
-import { formatUnits } from 'viem'
-import { PEANUT_WALLET_TOKEN_DECIMALS } from '@/constants/zerodev.consts'
-import { PostSignupActionManager } from '@/components/Global/PostSignupActionManager'
-import { useWithdrawFlow } from '@/context/WithdrawFlowContext'
-import { useClaimBankFlow } from '@/context/ClaimBankFlowContext'
-import { useNotifications } from '@/hooks/useNotifications'
-import { useCapabilities } from '@/hooks/useCapabilities'
-import { useCardInfo } from '@/hooks/useCardInfo'
-import HomeCarouselCTA from '@/components/Home/HomeCarouselCTA'
-import CardLaunchCTA from '@/components/Home/CardLaunchCTA'
-import EnableAutoBalanceBanner from '@/components/Home/EnableAutoBalanceBanner'
-import InvitesIcon from '@/components/Home/InvitesIcon'
-import NavigationArrow from '@/components/Global/NavigationArrow'
-import { updateUserById } from '@/app/actions/users'
-import { useHaptic } from 'use-haptic'
-import { useTranslations } from 'next-intl'
-import { useActivationStatus } from '@/hooks/useActivationStatus'
-import ActivationCTAs from '@/components/Home/ActivationCTAs'
-import PendingVerificationTasks from '@/components/Home/PendingVerificationTasks'
-import LazyLoadErrorBoundary from '@/components/Global/LazyLoadErrorBoundary'
-import posthog from 'posthog-js'
-import { ANALYTICS_EVENTS } from '@/constants/analytics.consts'
-import { MIGRATION_SURFACES } from '@/constants/migration.consts'
-import { useModalsContext } from '@/context/ModalsContext'
-
-// Lazy load heavy modal components (~20-30KB each) to reduce initial bundle size
-// Components are only loaded when user triggers them
-// Wrapped in error boundaries to gracefully handle chunk load failures
-const BalanceWarningModal = lazy(() => import('@/components/Global/BalanceWarningModal'))
-const SetupNotificationsModal = lazy(() => import('@/components/Notifications/SetupNotificationsModal'))
-const NoMoreJailModal = lazy(() => import('@/components/Global/NoMoreJailModal'))
-const EarlyUserModal = lazy(() => import('@/components/Global/EarlyUserModal'))
-const WelcomeUnlockModal = lazy(() => import('@/components/Home/WelcomeUnlockModal'))
-const IosPwaInstallModal = lazy(() => import('@/components/Global/IosPwaInstallModal'))
-const MigrationDownloadModal = lazy(() => import('@/components/Migration/MigrationDownloadModal'))
-const ScanToDownloadModal = lazy(() => import('@/components/Migration/ScanToDownloadModal'))
-const ReviewPromptModal = lazy(() => import('@/components/Migration/ReviewPromptModal'))
-
-const BALANCE_WARNING_THRESHOLD = parseInt(process.env.NEXT_PUBLIC_BALANCE_WARNING_THRESHOLD ?? '500')
-const BALANCE_WARNING_EXPIRY = parseInt(process.env.NEXT_PUBLIC_BALANCE_WARNING_EXPIRY ?? '1814400') // 21 days in seconds
+import { HomePage } from '@/features/home/HomePage'
 
 export default function Home() {
-    const t = useTranslations('home')
-    const tNav = useTranslations('navigation')
-    const { showPermissionModal } = useNotifications()
-    const { isGetAppModalOpen, setIsGetAppModalOpen } = useModalsContext()
-    const { balance, isFetchingBalance, spendableBalance, isFetchingSpendableBalance } = useWallet()
-    const { resetFlow: resetClaimBankFlow } = useClaimBankFlow()
-    const { resetWithdrawFlow } = useWithdrawFlow()
-    const { user } = useUserStore()
-    const [isBalanceHidden, setIsBalanceHidden] = useState(() => {
-        const prefs = user ? getUserPreferences(user.user.userId) : undefined
-        return prefs?.balanceHidden ?? false
-    })
-    const { isConnected: isWagmiConnected } = useAccount()
-    const { disconnect: disconnectWagmi } = useDisconnect()
-    const { triggerHaptic } = useHaptic()
-
-    const { isFetchingUser, fetchUser } = useAuth()
-    const { isKycApproved } = useCapabilities()
-    const { isActivated, activationStep, dismissCardStep } = useActivationStatus()
-    // Fire-and-forget: warms the card-info cache so /card mounts fast.
-    // Return values intentionally unused — only the fetch side effect matters.
-    useCardInfo()
-    const username = user?.user.username
-
-    const [showBalanceWarningModal, setShowBalanceWarningModal] = useState(false)
-    const [isPostSignupActionModalVisible, setIsPostSignupActionModalVisible] = useState(false)
-    const [showKycModal, setShowKycModal] = useState(false)
-    // migration download prompt outranks every other home modal (self-gating,
-    // only during the pwa-sunset notice window)
-    const [showMigrationModal, setShowMigrationModal] = useState(false)
-
-    // Track if this is a fresh signup session - captured once on mount so it persists
-    // even after NoMoreJailModal clears the sessionStorage key
-    const [_isPostSignupSession] = useState(() => {
-        if (typeof window === 'undefined') return false
-        return sessionStorage.getItem('showNoMoreJailModal') === 'true'
-    })
-
-    // re-fetch user on mount to pick up activation status changes (e.g. after QR payment)
-    useEffect(() => {
-        fetchUser()
-    }, []) // eslint-disable-line react-hooks/exhaustive-deps
-
-    // the migration prompt outranks the post-signup modal; unmounting the
-    // manager skips its onVisibilityChange(false), so clear the state here or
-    // it stays stuck true and suppresses the balance-warning/review modals
-    useEffect(() => {
-        if (showMigrationModal) setIsPostSignupActionModalVisible(false)
-    }, [showMigrationModal])
-
-    // Show the "You're unlocked" celebration exactly once: the user has a usable
-    // rail (isKycApproved) and has never dismissed it (activationCelebratedAt is
-    // null, stamped server-side on dismiss). A KYC re-approval can't resurface it
-    // — unlike the old showKycCompletedModal flag, which re-fired on every
-    // `→ approved` transition (e.g. the faster_payments endorsement backfill).
-    useEffect(() => {
-        if (isKycApproved && !user?.user.activationCelebratedAt) {
-            setShowKycModal(true)
-        }
-    }, [isKycApproved, user?.user.activationCelebratedAt])
-
-    const userFullName = useMemo(() => {
-        if (!user) return
-        return user.user.fullName
-    }, [user])
-
-    const handleToggleBalanceVisibility = useCallback(
-        (e: React.MouseEvent<HTMLButtonElement>) => {
-            e.stopPropagation()
-            setIsBalanceHidden((prev: boolean) => {
-                const newValue = !prev
-                posthog.capture(ANALYTICS_EVENTS.BALANCE_VISIBILITY_TOGGLED, { is_hidden: newValue })
-                if (user) {
-                    updateUserPreferences(user.user.userId, { balanceHidden: newValue })
-                }
-                return newValue
-            })
-        },
-        [user]
-    )
-
-    const isLoading = isFetchingUser && !username
-
-    useEffect(() => {
-        resetClaimBankFlow()
-        resetWithdrawFlow()
-    }, [resetClaimBankFlow, resetWithdrawFlow])
-
-    // always reset external wallet connection on home page
-    useEffect(() => {
-        if (isWagmiConnected) {
-            disconnectWagmi()
-        }
-    }, [isWagmiConnected, disconnectWagmi])
-
-    // effect for showing balance warning modal
-    // modal priority order: notifications -> kyc -> post signup -> ios pwa -> balance warning
-    // this ensures users complete important actions before seeing the balance warning
-    useEffect(() => {
-        if (isFetchingBalance || balance === undefined || !user) return
-
-        if (typeof window !== 'undefined') {
-            const userPreferences = getUserPreferences(user.user.userId)
-            const hasSeenBalanceWarning =
-                (userPreferences?.hasSeenBalanceWarning?.expiry ?? 0) > Date.now() &&
-                userPreferences?.hasSeenBalanceWarning?.value
-            const balanceInUsd = Number(formatUnits(balance, PEANUT_WALLET_TOKEN_DECIMALS))
-
-            // show if:
-            // 1. balance is above the threshold
-            // 2. user hasn't seen this warning in the current session
-            // 3. no other modals are currently active (respects priority order)
-            if (
-                balanceInUsd > BALANCE_WARNING_THRESHOLD &&
-                !hasSeenBalanceWarning &&
-                !showMigrationModal && // highest priority
-                !showPermissionModal &&
-                !showKycModal &&
-                !isPostSignupActionModalVisible
-            ) {
-                setShowBalanceWarningModal(true)
-            }
-        }
-    }, [
-        balance,
-        isFetchingBalance,
-        showMigrationModal,
-        showPermissionModal,
-        showKycModal,
-        isPostSignupActionModalVisible,
-        user,
-    ])
-
-    if (isLoading) {
-        return <Loading variant="mascot" coverFullScreen />
-    }
-
-    return (
-        <PageContainer>
-            <div className="space-y-6 h-full w-full p-5">
-                <div className="flex items-center justify-between gap-2">
-                    <UserHeader username={username!} fullName={userFullName} isVerified={isKycApproved} />
-                    {isActivated && (
-                        <Link onClick={() => triggerHaptic()} href="/rewards" className="flex items-center gap-0">
-                            <InvitesIcon />
-                            <span className="pl-1 text-sm font-semibold whitespace-nowrap md:text-base">
-                                {t('rewards')}
-                            </span>
-                            <NavigationArrow size={16} className="fill-black" />
-                        </Link>
-                    )}
-                    {/* <NotificationNavigation /> */}
-                </div>
-                <div className="space-y-4">
-                    <ActionButtonGroup>
-                        <ActionButtonWithHref label={tNav('add')} action="add" href="/add-money" size="small" />
-                        <ActionButtonWithHref
-                            label={tNav('withdraw')}
-                            action="withdraw"
-                            href="/withdraw"
-                            size="small"
-                        />
-                    </ActionButtonGroup>
-
-                    <WalletBalance
-                        balance={spendableBalance}
-                        isBalanceHidden={isBalanceHidden}
-                        onToggleBalanceVisibility={handleToggleBalanceVisibility}
-                        isFetchingBalance={isFetchingSpendableBalance}
-                    />
-
-                    <ActionButtonGroup>
-                        <ActionButtonWithHref
-                            label={tNav('send')}
-                            action="send"
-                            href="/send"
-                            variant="purple"
-                            size="large"
-                        />
-                        <ActionButtonWithHref
-                            label={tNav('request')}
-                            action="request"
-                            href="/request"
-                            variant="purple"
-                            size="large"
-                        />
-                    </ActionButtonGroup>
-                </div>
-
-                <div className="space-y-2">
-                    <EnableAutoBalanceBanner />
-                    {/* Public-launch splash. Self-gating (see CardLaunchCTA): shows post-
-                        launch to activated, geo-eligible, card-less, non-waitlisted users;
-                        dismiss/click hides it forever. Rendered above the carousel/activation
-                        CTAs so it leads the home stack on launch day. */}
-                    <CardLaunchCTA />
-                    {/* Pending Bridge verification tasks (ToS / hosted re-verification).
-                        Sibling of ActivationCTAs on purpose: it must show for users
-                        who can already transact (the advisory cohort), whom the
-                        activation card deliberately stands down for. Self-hiding;
-                        dismissible here — resurfaces under Profile → Unlocked regions. */}
-                    <PendingVerificationTasks dismissible />
-                    {isActivated ? (
-                        <HomeCarouselCTA />
-                    ) : (
-                        <ActivationCTAs activationStep={activationStep} onDismissCard={dismissCardStep} />
-                    )}
-                    <HomeHistory
-                        username={username ?? undefined}
-                        hideTxnAmount={isBalanceHidden}
-                        hideEmptyState={!isActivated}
-                    />
-                </div>
-
-                {showPermissionModal && !showBalanceWarningModal && !showMigrationModal && (
-                    <LazyLoadErrorBoundary>
-                        <Suspense fallback={null}>
-                            <SetupNotificationsModal />
-                        </Suspense>
-                    </LazyLoadErrorBoundary>
-                )}
-
-                <LazyLoadErrorBoundary>
-                    <Suspense fallback={null}>
-                        <MigrationDownloadModal onVisibilityChange={setShowMigrationModal} />
-                    </Suspense>
-                </LazyLoadErrorBoundary>
-
-                {/* desktop target of the get-the-app carousel CTA */}
-                {isGetAppModalOpen && (
-                    <LazyLoadErrorBoundary>
-                        <Suspense fallback={null}>
-                            <ScanToDownloadModal
-                                visible={isGetAppModalOpen}
-                                onClose={() => setIsGetAppModalOpen(false)}
-                                surface={MIGRATION_SURFACES.HOME_BANNER}
-                            />
-                        </Suspense>
-                    </LazyLoadErrorBoundary>
-                )}
-            </div>
-            {/* Add Money Prompt Modal */}
-            {/* TODO @dev Disabling this, re-enable after properly fixing */}
-            {/* <AddMoneyPromptModal visible={showAddMoneyPromptModal} onClose={() => setShowAddMoneyPromptModal(false)} /> */}
-
-            {/* these modals manage their own state internally */}
-            {!showBalanceWarningModal && !showMigrationModal && (
-                <>
-                    <LazyLoadErrorBoundary>
-                        <Suspense fallback={null}>
-                            <NoMoreJailModal />
-                        </Suspense>
-                    </LazyLoadErrorBoundary>
-
-                    <LazyLoadErrorBoundary>
-                        <Suspense fallback={null}>
-                            <EarlyUserModal />
-                        </Suspense>
-                    </LazyLoadErrorBoundary>
-                </>
-            )}
-
-            <LazyLoadErrorBoundary>
-                <Suspense fallback={null}>
-                    <WelcomeUnlockModal
-                        isOpen={showKycModal && !showBalanceWarningModal && !showMigrationModal}
-                        onClose={async () => {
-                            // close the modal immediately for better ux
-                            setShowKycModal(false)
-                            // update the database and refetch user to ensure sync
-                            if (user?.user.userId) {
-                                await updateUserById({
-                                    userId: user.user.userId,
-                                    dismissActivationCelebration: true,
-                                })
-                                // refetch user to ensure the modal doesn't reappear
-                                await fetchUser()
-                            }
-                        }}
-                    />
-                </Suspense>
-            </LazyLoadErrorBoundary>
-
-            {/* Balance Warning Modal */}
-            <LazyLoadErrorBoundary>
-                <Suspense fallback={null}>
-                    <BalanceWarningModal
-                        visible={showBalanceWarningModal && !showMigrationModal}
-                        onCloseAction={() => {
-                            setShowBalanceWarningModal(false)
-                            updateUserPreferences(user!.user.userId, {
-                                hasSeenBalanceWarning: {
-                                    value: true,
-                                    expiry: Date.now() + BALANCE_WARNING_EXPIRY * 1000,
-                                },
-                            })
-                        }}
-                    />
-                </Suspense>
-            </LazyLoadErrorBoundary>
-
-            <LazyLoadErrorBoundary>
-                <Suspense fallback={null}>
-                    <IosPwaInstallModal />
-                </Suspense>
-            </LazyLoadErrorBoundary>
-
-            {/* Card Pioneer Modal - Show to all users who haven't purchased */}
-            {/* Eligibility check happens during the flow (geo screen), not here */}
-            {/* unmounted while the migration prompt shows (it re-checks on
-                remount); the effect below clears its stuck visibility state */}
-            {!showMigrationModal && (
-                <PostSignupActionManager onActionModalVisibilityChange={setIsPostSignupActionModalVisible} />
-            )}
-
-            {/* App review nudge (native only, once ever) — lowest priority */}
-            {!showMigrationModal &&
-                !showPermissionModal &&
-                !showBalanceWarningModal &&
-                !showKycModal &&
-                !isPostSignupActionModalVisible && (
-                    <LazyLoadErrorBoundary>
-                        <Suspense fallback={null}>
-                            <ReviewPromptModal />
-                        </Suspense>
-                    </LazyLoadErrorBoundary>
-                )}
-        </PageContainer>
-    )
-}
-
-function WalletBalance({
-    balance,
-    isBalanceHidden,
-    onToggleBalanceVisibility,
-    isFetchingBalance,
-}: {
-    balance: bigint | undefined
-    isBalanceHidden: boolean
-    onToggleBalanceVisibility: (e: React.MouseEvent<HTMLButtonElement>) => void
-    isFetchingBalance?: boolean
-}) {
-    const balanceDisplay = useMemo(() => {
-        if (isBalanceHidden) {
-            return (
-                <span className="inline-flex items-center">
-                    <span className="relative top-1 text-[48px]">****</span>
-                </span>
-            )
-        }
-        return balance !== undefined ? formatExtendedNumber(printableUsdc(balance)) : ''
-    }, [isBalanceHidden, balance])
-
-    return (
-        <div className="flex items-center gap-4">
-            <div className="flex items-end gap-2 text-[48px] leading-none font-black md:text-[56px]">
-                {isFetchingBalance || balance === undefined ? (
-                    <span className="block pl-3">
-                        <Loading />
-                    </span>
-                ) : (
-                    <>
-                        <span className="text-[32px] md:text-[40px]">$ </span>
-                        {balanceDisplay}
-                    </>
-                )}
-            </div>
-
-            {!isFetchingBalance && (
-                <button onClick={onToggleBalanceVisibility}>
-                    <Icon name={isBalanceHidden ? 'eye-slash' : 'eye'} size={24} fill={'black'} />
-                </button> // no balance <> no icon
-            )}
-        </div>
-    )
-}
-
-interface ActionButtonProps {
-    label: string
-    action: 'add' | 'withdraw' | 'send' | 'request'
-    href: string
-    variant?: ButtonVariant
-    size?: ButtonSize
-}
-
-function ActionButtonWithHref({ label, action, href, variant = 'primary-soft', size = 'small' }: ActionButtonProps) {
-    return (
-        <Link href={href} className="block">
-            <ActionButton label={label} action={action} variant={variant} size={size} />
-        </Link>
-    )
-}
-
-function ActionButton({ label, action, variant = 'primary-soft', size = 'small' }: Omit<ActionButtonProps, 'href'>) {
-    const iconSize = size === 'large' ? 18 : 16
-    const renderIcon = (): React.ReactNode => {
-        switch (action) {
-            case 'send':
-                return <Icon name="arrow-up-right" size={iconSize} fill="currentColor" />
-            case 'withdraw':
-                return <Icon name="arrow-up" size={iconSize} fill="currentColor" />
-            case 'add':
-                return <Icon name="arrow-down" size={iconSize} fill="currentColor" />
-            case 'request':
-                return <Icon name="arrow-down-left" size={iconSize} fill="currentColor" />
-            default:
-                return null
-        }
-    }
-
-    return (
-        <Button
-            variant={variant}
-            className={twMerge(
-                'flex w-auto cursor-pointer items-center justify-center rounded-full',
-                size === 'large'
-                    ? 'h-12 gap-x-2 px-6 md:h-14 md:px-7' // Send/Request size
-                    : 'h-10 gap-x-1 px-5 md:h-12 md:px-6' // Add/Withdraw size
-            )}
-            shadowSize="4"
-            size={size}
-        >
-            {renderIcon()}
-            <span
-                className={twMerge(
-                    'font-semibold whitespace-nowrap',
-                    size === 'small'
-                        ? 'text-sm md:text-base' // Add/Withdraw size
-                        : 'text-base md:text-lg' // Send/Request size
-                )}
-            >
-                {label}
-            </span>
-        </Button>
-    )
-}
-
-function ActionButtonGroup({ children }: { children: React.ReactNode }) {
-    return <div className="flex items-center justify-normal gap-4">{children}</div>
+    return <HomePage />
 }
